+++ src/server/devices/IoBrokerDeviceInfo.ts
@@ -9,8 +9,6 @@ export class IoBrokerDeviceInfo extends DeviceInfo {
   public deviceRoomIndex: number;
   public type: 'device' | 'channel' | 'state';
   public fullID: string;
-  public channel?: number;
-  public valueName?: string;
   public devConf: deviceConfig;
 
   public static idSplitter(id: string): string[] {
@@ -28,64 +26,72 @@ export class IoBrokerDeviceInfo extends DeviceInfo {
   /**
    * Extracts the relevant infos from the passed deviceConfig and combines them in a new Info object
    * @param pDevConf - The device Config based on the extracted devices.json from ioBroker
-   * @param isJsStateChildObject - Within JS Objects, creating devices is limited,
-   * so we name the first child state for the object creation (e.g. javascript.0.00-EnergyManager.CurrentProduction)
+   * @param deviceId - The id of the device
+   * @param deviceType - The type of the device
+   * @param room - The room id of the device
+   * @param deviceRoomIndex - Index of this device in regards to the devicetype.
    */
-  public constructor(pDevConf: deviceConfig, isJsStateChildObject: boolean = false) {
+  public constructor(
+    pDevConf: deviceConfig,
+    deviceId: string,
+    deviceType: string,
+    room: string,
+    deviceRoomIndex: number,
+  ) {
     super();
     this.devConf = pDevConf;
     this.type = pDevConf.type as 'device' | 'channel' | 'state';
-
-    const idSplit: string[] = IoBrokerDeviceInfo.idSplitter(pDevConf._id);
     this.fullID = pDevConf._id;
-    this.devID = idSplit[2];
+    this.devID = deviceId;
     this.fullName = pDevConf.common!.name as string;
+    this.deviceType = deviceType;
+    this.deviceRoomIndex = deviceRoomIndex;
+    this.room = room;
+  }
+
+  public static byStateJsSplit(pDevConf: deviceConfig): IoBrokerDeviceInfo {
     const nameSplit: string[] = (pDevConf.common!.name as string).split('-');
+    const idSplit: string[] = IoBrokerDeviceInfo.idSplitter(pDevConf._id);
+    /**
+     * Name-Split
+     * 0: Indikator own "00"
+     * 1: "EnergyManager"
+     * 2: Raum
+     * 3: Was für ein Gerät
+     * 4: Index dieses Gerätes im Raum (ggf. + :Channel)
+     * 5?: Name des Wertes
+     */
+    const deviceType = nameSplit[1];
+    const room = nameSplit.length >= 3 ? nameSplit[2] : '';
+    const deviceRoomIndex = nameSplit.length >= 4 ? Number(nameSplit[3]) : 0;
+    return new IoBrokerDeviceInfo(pDevConf, idSplit[2], deviceType, room, deviceRoomIndex);
+  }
 
-    if (!isJsStateChildObject) {
-      /**
-       * 0: hm-rpc
-       * 1: rcpInstance
-       * 2: Device ID
-       * 3?: Channel
-       * 4?: ValueName
-       */
+  public static byDeviceConfig(pDevConf: deviceConfig): IoBrokerDeviceInfo {
+    const nameSplit: string[] = (pDevConf.common!.name as string).split('-');
+    const idSplit: string[] = IoBrokerDeviceInfo.idSplitter(pDevConf._id);
+    let channel: number | undefined;
+    /**
+     * 0: hm-rpc
+     * 1: rcpInstance
+     * 2: Device ID
+     * 3?: Channel
+     * 4?: ValueName
+     */
+    /**
+     * Name-Split
+     * 0: Indikator own "00"
+     * 1: "HmIP"
+     * 2: Raum
+     * 3: Was für ein Gerät
+     * 4: Index dieses Gerätes im Raum (ggf. + : Channel)
+     * 5?: Name des Wertes
+     */
 
-      if (idSplit.length > 3) {
-        this.channel = Number(idSplit[3]);
-      }
-
-      if (idSplit.length > 4) {
-        this.valueName = idSplit[4];
-      }
-      /**
-       * Name-Split
-       * 0: Indikator own "00"
-       * 1: "HmIP"
-       * 2: Raum
-       * 3: Was für ein Gerät
-       * 4: Index dieses Gerätes im Raum (ggf. + :Channel)
-       * 5?: Name des Wertes
-       */
-
-      this.room = nameSplit[2];
-      this.deviceType = nameSplit[3];
-      this.deviceRoomIndex = Number(nameSplit[4].split(':')[0]);
-      return;
-    } else {
-      /**
-       * Name-Split
-       * 0: Indikator own "00"
-       * 1: "EnergyManager"
-       * 2: Raum
-       * 3: Was für ein Gerät
-       * 4: Index dieses Gerätes im Raum (ggf. + :Channel)
-       * 5?: Name des Wertes
-       */
-      this.deviceType = nameSplit[1];
-      this.room = nameSplit.length >= 3 ? nameSplit[2] : '';
-      this.deviceRoomIndex = nameSplit.length >= 4 ? Number(nameSplit[3]) : 0;
-    }
+    const room = nameSplit[2];
+    const deviceType = nameSplit[3];
+    const deviceRoomIndex: number = Number(nameSplit[4].split(':')[0]);
+    return new IoBrokerDeviceInfo(pDevConf, idSplit[2], deviceType, room, deviceRoomIndex);
   }
 
   public override toJSON(): Partial<IoBrokerDeviceInfo> {