+++ src/server/devices/smartGarden/index.ts
@@ -1,0 +1,2 @@
+export * from './SmartGardenService';
+export * from './SmartGardenDeviceRegistrationInfo';