+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Preparations for the implementation of SmartGardenDevices
+
 ## 3.0.0-alpha.61 (2024-05-10)
 
 * (theimo1221) Fix minor logic error within dachs